Organize sample code and comment combine channels screen
